Break up render into modules.
