+++ battleground_viewer/src/construct_render/render/batching.rs
@@ -1,0 +1,132 @@
+use three_d::*;
+
+use battleground_construct::display::primitives::Primitive;
+
+pub trait BatchKey {
+    fn to_batch_key(&self) -> u64;
+}
+
+impl BatchKey for Primitive {
+    fn to_batch_key(&self) -> u64 {
+        use std::hash::Hash;
+        use std::hash::Hasher;
+        let mut hasher = std::collections::hash_map::DefaultHasher::new();
+        let state = &mut hasher;
+        match *self {
+            Primitive::Cuboid(cube) => {
+                0usize.hash(state);
+                cube.width.to_bits().hash(state);
+                cube.length.to_bits().hash(state);
+                cube.height.to_bits().hash(state);
+            }
+            Primitive::Sphere(sphere) => {
+                1usize.hash(state);
+                sphere.radius.to_bits().hash(state);
+            }
+            Primitive::Cylinder(cylinder) => {
+                2usize.hash(state);
+                cylinder.radius.to_bits().hash(state);
+                cylinder.height.to_bits().hash(state);
+            }
+            Primitive::Line(_line) => {
+                // All lines hash the same.
+                3usize.hash(state);
+            }
+            Primitive::Cone(cone) => {
+                4usize.hash(state);
+                cone.radius.to_bits().hash(state);
+                cone.height.to_bits().hash(state);
+            }
+            Primitive::Circle(circle) => {
+                5usize.hash(state);
+                circle.radius.to_bits().hash(state);
+            }
+        }
+        // val
+        hasher.finish()
+    }
+}
+
+#[derive(Debug, Copy, Clone)]
+pub enum BatchProperties {
+    None,
+    Basic { is_transparent: bool },
+}
+
+impl BatchKey for BatchProperties {
+    fn to_batch_key(&self) -> u64 {
+        use std::hash::Hash;
+        use std::hash::Hasher;
+        let mut hasher = std::collections::hash_map::DefaultHasher::new();
+        let state = &mut hasher;
+        match *self {
+            BatchProperties::None => {
+                1usize.hash(state);
+            }
+            BatchProperties::Basic { is_transparent } => {
+                2usize.hash(state);
+                is_transparent.hash(state);
+            }
+        }
+        hasher.finish()
+    }
+}
+
+/// Produces material values based on batch property
+pub trait BatchMaterial {
+    fn new_for_batch(context: &Context, batch_properties: BatchProperties) -> Self
+    where
+        Self: Material;
+}
+
+impl BatchMaterial for PhysicalMaterial {
+    fn new_for_batch(context: &Context, batch_properties: BatchProperties) -> PhysicalMaterial {
+        let material_new = match batch_properties {
+            BatchProperties::None => PhysicalMaterial::new_opaque,
+            BatchProperties::Basic {
+                is_transparent: false,
+            } => PhysicalMaterial::new_opaque,
+            BatchProperties::Basic {
+                is_transparent: true,
+            } => PhysicalMaterial::new_transparent,
+        };
+        material_new(
+            context,
+            &CpuMaterial {
+                albedo: Color {
+                    r: 255,
+                    g: 255,
+                    b: 255,
+                    a: 255,
+                },
+                ..Default::default()
+            },
+        )
+    }
+}
+
+impl BatchMaterial for ColorMaterial {
+    fn new_for_batch(context: &Context, batch_properties: BatchProperties) -> ColorMaterial {
+        let material_new = match batch_properties {
+            BatchProperties::None => ColorMaterial::new_opaque,
+            BatchProperties::Basic {
+                is_transparent: false,
+            } => ColorMaterial::new_opaque,
+            BatchProperties::Basic {
+                is_transparent: true,
+            } => ColorMaterial::new_transparent,
+        };
+        material_new(
+            context,
+            &CpuMaterial {
+                albedo: Color {
+                    r: 255,
+                    g: 255,
+                    b: 255,
+                    a: 255,
+                },
+                ..Default::default()
+            },
+        )
+    }
+}
